+++ lib/domain/diary_setting_physical_condtion_detail/mutation.dart
@@ -1,6 +1,6 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:pilll/database/database.dart';
+import 'package:pilll/provider/database.dart';
 import 'package:pilll/entity/diary_setting.codegen.dart';
 import 'package:pilll/util/datetime/day.dart';
