+++ lib/provider/diary.dart
@@ -1,5 +1,5 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:pilll/database/database.dart';
+import 'package:pilll/provider/database.dart';
 import 'package:pilll/entity/diary.codegen.dart';
 import 'package:riverpod/riverpod.dart';
 import 'package:pilll/util/datetime/day.dart';
