+++ lib/provider/diary_setting.dart
@@ -1,5 +1,5 @@
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:pilll/database/database.dart';
+import 'package:pilll/provider/database.dart';
 import 'package:pilll/entity/diary_setting.codegen.dart';
 
 final diarySettingProvider =
