+++ test/domain/record/record_page_button_state_test.dart
@@ -1,5 +1,5 @@
 import 'package:pilll/analytics.dart';
-import 'package:pilll/database/database.dart';
+import 'package:pilll/provider/database.dart';
 import 'package:pilll/domain/record/components/button/cancel_button.dart';
 import 'package:pilll/domain/record/components/button/record_page_button.dart';
 import 'package:pilll/domain/record/components/button/taken_button.dart';
